c++: Make __builtin_launder reject invalid types [PR116673]
The standard says that std::launder is ill-formed for function pointers
and cv void pointers, so there's no reason for __builtin_launder to
accept them. This change allows implementations of std::launder to defer
to the built-in for error checking, although libstdc++ will continue to
diagnose it directly for more user-friendly diagnostics.

	PR c++/116673

gcc/cp/ChangeLog:

	* semantics.cc (finish_builtin_launder): Diagnose function
	pointers and cv void pointers.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1z/launder2.C: Adjust dg-error strings.
	* g++.dg/cpp1z/launder10.C: New test.