c++/modules: Allow imported references in constant expressions
Currently the streaming code uses TREE_CONSTANT to determine whether an
entity will have a definition that is interesting to stream out.  This
is not sufficient, however; we also need to write the definition of
references, since although not TREE_CONSTANT they can still be usable in
constant expressions.

As such this patch uses the existing decl_maybe_constant_var function
which correctly handles this case.

gcc/cp/ChangeLog:

	* module.cc (has_definition): Use decl_maybe_constant_var
	instead of TREE_CONSTANT.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/cexpr-5_a.C: New test.
	* g++.dg/modules/cexpr-5_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>
Reviewed-by: Jason Merrill <jason@redhat.com>