arm: testsuite: make use of -mcpu=unset/-march=unset
This patch makes use of the new ability to unset the CPU or
architecture flags on the command line to enable several more tests on
Arm.  It doesn't cover every case and it does enable some tests that
now fail for different reasons when the tests are no-longer skipped;
these were failing anyway for other testsuite configurations, so it's
still an overall improvement.

There's some restructuring required to fully implement this change: we
could previously treat Xscale as an architecture, even though the
option set -mcpu=, we now need to handle this correctly so that we
unset the architecture rather than the CPU.  To do this I've added a
new table for these variants and renamed the template functions to use
'cpu' rather than 'arch'.  This entailed updating the two XScale
related tests accordingly.

gcc/testsuite/ChangeLog:

	* lib/target-supports.exp: Move xscale to new generator table.
	(check_effective_target_arm_arch_FUNC_ok): Add -mcpu=unset to the
	list of flags.
	(add_options_for_arm_arch_FUNC): Likewise.
	(check_effective_target_arm_cpu_FUNC_ok): New function.
	(add_options_for_arm_cpu_FUNC): Likewise.
	(check_effective_target_arm_cpu_FUNC_link): Likewise.
	(check_effective_target_arm_cpu_FUNC_multilib): Likewise.
	* gcc.target/arm/g2.c: Update dg directives.
	* gcc.target/arm/scd42-2.c: Likewise.