ada: Pass parameters of full access unconstrained array types by copy in calls
When a full access array type is declared, either Volatile_Full_Access in
Ada 2012 or Atomic in Ada 2022, an implicit base array type is built by the
compiler with the Full_Access flag set, although full accesses cannot be
generated for objects of this type because the size is not known statically.

If the component type is a record with default values, an initialization
procedure taking a parameter of the base array type is built. Given that
full accesses cannot be generated for the parameter inside the procedure,
we need to pass the actual parameter by copy to the procedure in order to
implement the full access semantics.

gcc/ada/ChangeLog:

	* exp_ch6.adb (Expand_Actuals.Is_Legal_Copy): Return True for an
	initialization procedure with a full access formal parameter.
	(Expand_Actuals.Requires_Atomic_Or_Volatile_Copy): Return True if
	the formal parameter is of a full access unconstrained array type.