phi-opt: Improve heuristics for factoring out with constant (again) [PR116699]
The heuristics for factoring out with a constant checks that the assignment statement
is the last statement of the basic block but sometimes there is a predicate or a nop statement
after the assignment. Rejecting this case does not make sense since both predicates and nop
statements are removed and don't contribute any instructions. So we should skip over them
when checking if the assignment statement was the last statement in the basic block.

phi-opt-factor-1.c's f0 is such an example where it should catch it at phiopt1 (before predicates are removed)
and should happen in a similar way as f1 (which uses a temporary variable rather than return).

Bootstrapped and tested on x86_64-linux-gnu.

	PR tree-optimization/116699

gcc/ChangeLog:

	* tree-ssa-phiopt.cc (factor_out_conditional_operation): Skip over nop/predicates
	for seeing the assignment is the last statement.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/phi-opt-factor-1.c: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>